requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
line-bot-sdk==3.11.0
